pyproject.toml
[project]
name = "mathjax-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pillow>=12.0.0",
    "prompt-toolkit>=3.0.52",
    "pywin32>=311",
]
